+++ git-remote-updater.php
@@ -14,7 +14,7 @@
  * Description:       Allows you to easily update Git Updater repositories in bulk via REST API endpoint updating. Requires Git Updater PRO.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           2.3.1.2
+ * Version:           2.3.1.3
  * License:           MIT
  * Network:           true
  * Domain Path:       /languages
@@ -38,10 +38,20 @@ if ( ! defined( 'WPINC' ) ) {
 // Load Autoloader.
 require_once __DIR__ . '/vendor/autoload.php';
 
-( new GRU_Freemius() )->init();
+// Make sure `is_plugin_active()` is available.
+if ( ! function_exists( 'is_plugin_active' ) ) {
+	require_once ABSPATH . '/wp-admin/includes/plugin.php';
+}
+
+// Don't load if Git Updater not running.
+if ( ! is_plugin_active( 'git-updater/git-updater.php' ) ) {
+	return;
+}
+
 add_action(
 	'plugins_loaded',
 	function() {
+		( new GRU_Freemius() )->init();
 		( new Bootstrap() )->run();
 	}
 );